[#1382] neofs-cli: move `netmap` command to a separate package
